requirements.txt
Django==1.10.4
wheel==0.24.0
djangorestframework==3.3.3
